EEP-NutritionController/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
requests==2.31.0
boto3==1.28.15
s3fs==2023.6.0
joblib==1.4.2
scikit-learn==1.6.1
psycopg2>=2.9.1
python-dotenv>=0.19.0
bcrypt>=4.0.1
prometheus_client==0.19.0